pyproject.toml
## Project requirements (PEP 518, https://python-poetry.org/).


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]


[tool.black]
line-length = 96
target-version = ["py39", "py310", "py311", "py312"]


[tool.isort]
line_length = 96
profile = "black"


[tool.mypy]
explicit_package_bases = true
namespace_packages = true

check_untyped_defs = true
strict_optional = true
no_implicit_optional = true

warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true

color_output = false
pretty = true
show_error_codes = true

incremental = true
sqlite_cache = true


[tool.poetry]
name = "macrostrat_nbmg_import"
version = "0.0.1"
description = "Scripts for importing maps from NBMG"
license = "Apache-2.0"
authors = [
    "Cannon Lock <clock@wisc.edu>",
    "Brian Aydemir <baydemir@morgridge.org>",
]
repository = "https://github.com/brianaydemir/criticalmaas-nbmg-import"
packages = [{ include = "macrostrat/nbmg_import" }]


[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "~23.2"
aiohttp = "~3.9"
anyio = "~4.2"
beautifulsoup4 = "~4.12"
minio = "~7.2"
requests = "~2.31"
psycopg = {extras = ["binary"], version = "~3.1"}
python-dotenv = "~1.0"
python-magic = "~0.4"
"macrostrat.map_integration" = { path = "vendor/macrostrat/map-integration", develop = true }


[tool.poetry.group.dev.dependencies]
bandit = "~1.7"
black = "~23.12"
docutils = "~0.20"
isort = "~5.13"
mypy = "~1.8"
pre-commit = "~3.6"
pylint = "~3.0"
safety = "~2.3"
types-beautifulsoup4 = "~4.12"
types-requests = "~2.31"


[tool.pylint.messages-control]
max-line-length = 96
disable = "invalid-name,missing-class-docstring,missing-function-docstring"
